+++ hutool-crypto/src/main/java/cn/hutool/crypto/asymmetric/AsymmetricCrypto.java
@@ -1,6 +1,7 @@
 package cn.hutool.crypto.asymmetric;
 
 import java.io.InputStream;
+import java.nio.charset.Charset;
 import java.security.PrivateKey;
 import java.security.PublicKey;
 
@@ -10,6 +11,7 @@ import cn.hutool.core.codec.Base64;
 import cn.hutool.core.io.IORuntimeException;
 import cn.hutool.core.io.IoUtil;
 import cn.hutool.core.util.CharsetUtil;
+import cn.hutool.core.util.HexUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.crypto.CryptoException;
 import cn.hutool.crypto.symmetric.SymmetricAlgorithm;
@@ -18,35 +20,38 @@ import cn.hutool.crypto.symmetric.SymmetricAlgorithm;
  * 非对称加密算法<br>
  * 1、签名：使用私钥加密，公钥解密。用于让所有公钥所有者验证私钥所有者的身份并且用来防止私钥所有者发布的内容被篡改，但是不用来保证内容不被他人获得。<br>
  * 2、加密：用公钥加密，私钥解密。用于向公钥所有者发布信息,这个信息可能被他人篡改,但是无法被他人获得。
+ * 
  * @author Looly
  *
  */
-public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
+public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto> {
 
 	/** Cipher负责完成加密或解密工作 */
 	protected Cipher clipher;
-	
+
 	// ------------------------------------------------------------------ Constructor start
 	/**
 	 * 构造，创建新的私钥公钥对
+	 * 
 	 * @param algorithm {@link SymmetricAlgorithm}
 	 */
 	public AsymmetricCrypto(AsymmetricAlgorithm algorithm) {
-		this(algorithm, (byte[])null, (byte[])null);
+		this(algorithm, (byte[]) null, (byte[]) null);
 	}
-	
+
 	/**
 	 * 构造，创建新的私钥公钥对
+	 * 
 	 * @param algorithm 算法
 	 */
 	public AsymmetricCrypto(String algorithm) {
-		this(algorithm, (byte[])null, (byte[])null);
+		this(algorithm, (byte[]) null, (byte[]) null);
 	}
-	
+
 	/**
-	 * 构造
-	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
+	 * 构造 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
 	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
+	 * 
 	 * @param algorithm {@link SymmetricAlgorithm}
 	 * @param privateKeyBase64 私钥Base64
 	 * @param publicKeyBase64 公钥Base64
@@ -54,11 +59,11 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 	public AsymmetricCrypto(AsymmetricAlgorithm algorithm, String privateKeyBase64, String publicKeyBase64) {
 		this(algorithm.getValue(), Base64.decode(privateKeyBase64), Base64.decode(publicKeyBase64));
 	}
-	
+
 	/**
-	 * 构造
-	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
+	 * 构造 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
 	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
+	 * 
 	 * @param algorithm {@link SymmetricAlgorithm}
 	 * @param privateKey 私钥
 	 * @param publicKey 公钥
@@ -66,11 +71,11 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 	public AsymmetricCrypto(AsymmetricAlgorithm algorithm, byte[] privateKey, byte[] publicKey) {
 		this(algorithm.getValue(), privateKey, publicKey);
 	}
-	
+
 	/**
-	 * 构造
-	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
+	 * 构造 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
 	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
+	 * 
 	 * @param algorithm {@link SymmetricAlgorithm}
 	 * @param privateKey 私钥
 	 * @param publicKey 公钥
@@ -79,11 +84,11 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 	public AsymmetricCrypto(AsymmetricAlgorithm algorithm, PrivateKey privateKey, PublicKey publicKey) {
 		this(algorithm.getValue(), privateKey, publicKey);
 	}
-	
+
 	/**
-	 * 构造
-	 * 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
+	 * 构造 私钥和公钥同时为空时生成一对新的私钥和公钥<br>
 	 * 私钥和公钥可以单独传入一个，如此则只能使用此钥匙来做加密或者解密
+	 * 
 	 * @param algorithm 非对称加密算法
 	 * @param privateKeyBase64 私钥Base64
 	 * @param publicKeyBase64 公钥Base64
@@ -105,7 +110,7 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 	public AsymmetricCrypto(String algorithm, byte[] privateKey, byte[] publicKey) {
 		super(algorithm, privateKey, publicKey);
 	}
-	
+
 	/**
 	 * 构造
 	 * 
@@ -139,11 +144,11 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 		} catch (Exception e) {
 			throw new CryptoException(e);
 		}
-		
+
 		super.init(algorithm, privateKey, publicKey);
 		return this;
 	}
-	
+
 	// --------------------------------------------------------------------------------- Encrypt
 	/**
 	 * 加密
@@ -166,6 +171,29 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 	}
 
 	/**
+	 * 编码为Hex字符串
+	 * 
+	 * @param data 被加密的bytes
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return Hex字符串
+	 */
+	public String encryptHex(byte[] data, KeyType keyType) {
+		return HexUtil.encodeHexStr(encrypt(data, keyType));
+	}
+
+	/**
+	 * 编码为Base64字符串
+	 * 
+	 * @param data 被加密的bytes
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return Base64字符串
+	 * @since 4.0.1
+	 */
+	public String encryptBase64(byte[] data, KeyType keyType) {
+		return Base64.encode(encrypt(data, keyType));
+	}
+
+	/**
 	 * 加密
 	 * 
 	 * @param data 被加密的字符串
@@ -174,6 +202,18 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 	 * @return 加密后的bytes
 	 */
 	public byte[] encrypt(String data, String charset, KeyType keyType) {
+		return encrypt(StrUtil.bytes(data, charset), keyType);
+	}
+
+	/**
+	 * 加密
+	 * 
+	 * @param data 被加密的字符串
+	 * @param charset 编码
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return 加密后的bytes
+	 */
+	public byte[] encrypt(String data, Charset charset, KeyType keyType) {
 		return encrypt(StrUtil.bytes(data, charset), keyType);
 	}
 
@@ -189,15 +229,88 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 	}
 
 	/**
-	 * 加密
+	 * 编码为Hex字符串
 	 * 
 	 * @param data 被加密的字符串
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return Hex字符串
+	 * @since 4.0.1
+	 */
+	public String encryptHex(String data, KeyType keyType) {
+		return HexUtil.encodeHexStr(encrypt(data, keyType));
+	}
+
+	/**
+	 * 编码为Hex字符串
+	 * 
+	 * @param data 被加密的bytes
+	 * @param charset 编码
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return Hex字符串
+	 * @since 4.0.1
+	 */
+	public String encryptHex(String data, Charset charset, KeyType keyType) {
+		return HexUtil.encodeHexStr(encrypt(data, charset, keyType));
+	}
+
+	/**
+	 * 编码为Base64字符串
+	 * 
+	 * @param data 被加密的字符串
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return Base64字符串
+	 * @since 4.0.1
+	 */
+	public String encryptBase64(String data, KeyType keyType) {
+		return Base64.encode(encrypt(data, keyType));
+	}
+
+	/**
+	 * 编码为Base64字符串
+	 * 
+	 * @param data 被加密的字符串
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return Base64字符串
+	 * @since 4.0.1
+	 */
+	public String encryptBase64(String data, Charset charset, KeyType keyType) {
+		return Base64.encode(encrypt(data, charset, keyType));
+	}
+
+	/**
+	 * 加密
+	 * 
+	 * @param data 被加密的数据流
 	 * @param keyType 私钥或公钥 {@link KeyType}
 	 * @return 加密后的bytes
 	 * @throws IORuntimeException IO异常
 	 */
-	public byte[] encrypt(InputStream data, KeyType keyType) throws IORuntimeException{
+	public byte[] encrypt(InputStream data, KeyType keyType) throws IORuntimeException {
 		return encrypt(IoUtil.readBytes(data), keyType);
+	}
+
+	/**
+	 * 编码为Hex字符串
+	 * 
+	 * @param data 被加密的数据流
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return Hex字符串
+	 * @since 4.0.1
+	 */
+	public String encryptHex(InputStream data, KeyType keyType) {
+		return HexUtil.encodeHexStr(encrypt(data, keyType));
+	}
+
+	/**
+	 * 编码为Base64字符串
+	 * 
+	 * @param data 被加密的数据流
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return Base64字符串
+	 * @since 4.0.1
+	 */
+	public String encryptBase64(InputStream data, KeyType keyType) {
+		return Base64.encode(encrypt(data, keyType));
 	}
 
 	// --------------------------------------------------------------------------------- Decrypt
@@ -228,8 +341,32 @@ public class AsymmetricCrypto extends BaseAsymmetric<AsymmetricCrypto>{
 	 * @return 解密后的bytes
 	 * @throws IORuntimeException IO异常
 	 */
-	public byte[] decrypt(InputStream data, KeyType keyType) throws IORuntimeException{
+	public byte[] decrypt(InputStream data, KeyType keyType) throws IORuntimeException {
 		return decrypt(IoUtil.readBytes(data), keyType);
+	}
+
+	/**
+	 * 从Hex字符串解密
+	 * 
+	 * @param hexStr Hex字符串
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return 解密后的bytes
+	 * @since 4.0.1
+	 */
+	public byte[] decryptFromHex(String hexStr, KeyType keyType) {
+		return decrypt(HexUtil.decodeHex(hexStr), keyType);
+	}
+
+	/**
+	 * 从Base64字符串解密
+	 * 
+	 * @param hexStr Base64字符串
+	 * @param keyType 私钥或公钥 {@link KeyType}
+	 * @return 解密后的bytes
+	 * @since 4.0.1
+	 */
+	public byte[] decryptFromBase64(String base64Str, KeyType keyType) {
+		return decrypt(Base64.decode(base64Str, CharsetUtil.CHARSET_UTF_8), keyType);
 	}
 
 	// --------------------------------------------------------------------------------- Getters and Setters