+++ hutool-crypto/src/main/java/cn/hutool/crypto/digest/DigestUtil.java
@@ -361,4 +361,16 @@ public class DigestUtil {
 	public static HMac hmac(HmacAlgorithm algorithm, SecretKey key) {
 		return new HMac(algorithm, key);
 	}
+	
+	/**
+	 * 新建摘要器
+	 * 
+	 * @param algorithm 签名算法
+	 * @param data 数据
+	 * @return Digester
+	 * @since 4.0.1
+	 */
+	public static Digester digester(DigestAlgorithm algorithm) {
+		return new Digester(algorithm);
+	}
 }